refactor: split Swiss legal multilingual tasks into modular package
Move Swiss legal task definitions into prompts/metrics/main modules and keep backward compatibility via swiss_legal_evals re-export.
