+++ src/lighteval/tasks/multilingual/tasks/swiss_legal/__init__.py
@@ -1,0 +1,4 @@
+from lighteval.tasks.multilingual.tasks.swiss_legal.main import TASKS_TABLE
+
+
+__all__ = ["TASKS_TABLE"]
